requirements.txt
streamlit
pillow
numpy
matplotlib
opencv-python
